log_consumer/requirements.txt
# kafka-python==2.0.2
# psycopg2-binary==2.9.5

kafka-python==2.0.2
mysql-connector-python==8.2.0
